requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
json-rpc==1.10.8
python-dotenv==0.8.2
python-owasp-zap-v2.4==0.0.14
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
